WASP-1406/consolidate webhook variants into single webhook_template resource
Torq, Tines, and Coralogix were exposed as dedicated resources, but Orca
stores every one as a webhook variant (service_name="webhook") that differs
only by the config.type discriminator. The dedicated resources were pure
clones and duplicated the generic orcasecurity_integration_webhook_template,
which already accepts all types via its `type` argument.

- Delete the torq/tines/coralogix resource packages, their api_client
  wrappers, docs, templates, and examples.
- Fold the shared webhook helpers (header/body_fields conversion, plan
  modifiers, extract hooks) from webhook_variant_common into the webhook
  package; port their unit tests.
- Drop the three resources + imports from the provider registration.
- Document in the webhook_template page that a single resource covers
  common/torq/tines/opus/coralogix/panther via `type`, with a vendor example.
