--- templates/resources/integration_tines_template.md.tmpl
@@ -1,74 +1,0 @@
----
-# generated by https://github.com/hashicorp/terraform-plugin-docs
-page_title: "orcasecurity_integration_tines_template Resource - orcasecurity"
-subcategory: ""
-description: |-
-  Manage a Tines integration in Orca.
----
-
-# orcasecurity_integration_tines_template (Resource)
-
-Manages a [Tines](https://www.tines.com/) integration in Orca Security.
-
-Orca stores Tines as a webhook variant: under the hood this resource creates an
-external service config of `service_name = "webhook"` with `type = "tines"`.
-You could achieve the same result with `orcasecurity_integration_webhook_template`, but
-the dedicated resource hides the webhook plumbing and mirrors how the Orca UI
-displays the integration.
-
-## Example Usage
-
-### Basic example
-
-{{tffile "examples/resources/orcasecurity_integration_tines_template/resource.tf"}}
-
-### Sourcing the API key from HashiCorp Vault
-
-```terraform
-data "vault_generic_secret" "tines" {
-  path = "secret/tines/orca"
-}
-
-resource "orcasecurity_integration_tines_template" "from_vault" {
-  template_name = "tines-prod"
-  webhook_url   = "https://your-tenant.tines.com/webhook/<path>/<secret>"
-  api_key       = data.vault_generic_secret.tines.data["api_key"]
-}
-```
-
-## Argument Reference
-
-* `template_name` — (Required, String) Identifier for the integration as shown
-  in the Orca UI. Changing this forces a new resource.
-* `webhook_url` — (Required, String) Tines webhook URL Orca posts events to.
-* `api_key` — (Required, String, Sensitive) Tines API key sent with each
-  request.
-* `body_fields` — (Optional, List of String) Optional list of Orca alert fields
-  to include in the request body.
-* `custom_headers` — (Optional, Map of List of Object) Optional custom HTTP
-  headers, keyed by header name. Each value is a list of
-  `{ custom = "<value>" }` objects so a single header name can carry multiple
-  values.
-* `is_enabled` — (Optional, Bool) Defaults to `true`.
-* `is_default` — (Optional, Bool) Defaults to `false`.
-* `business_units` — (Optional, Set of String) Optional set of Orca business
-  unit IDs that may use this integration.
-
-## Attribute Reference
-
-* `id` — (String) Orca external service config identifier (UUID).
-
-## Notes
-
-* `api_key` lives inside the JSON config (not in SSM), so Orca's GET response
-  includes it. The provider still marks it as `sensitive = true` to keep it
-  out of CLI output and never overwrites the planned value during Create or
-  Update.
-* Use an encrypted remote backend (S3 + KMS, GCS, Terraform Cloud, etc.) and
-  restrict read access to the state.
-
-## Import
-
-```bash
-terraform import orcasecurity_integration_tines_template.example tines_name
-```
